# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Wind farm API Documentation",
        "description": "Wind farm API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/farms": {
            "get": {
                "summary": "List all farm details",
                "operationId": "cd8552f0d16d177a9f8645173cb0502f",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FarmCollection"
                                }
                            }
                        }
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerError"
                    }
                }
            }
        },
        "/api/farms/{farmId}": {
            "get": {
                "summary": "Show farm details",
                "operationId": "169767502ad0c493476ce5fa5ef2e72d",
                "parameters": [
                    {
                        "name": "farmId",
                        "in": "path",
                        "description": "The ID of the farm",
                        "required": true,
                        "example": 5
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FarmResource"
                                }
                            }
                        }
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerError"
                    }
                }
            }
        },
        "/api/turbines": {
            "get": {
                "summary": "List all turbines",
                "operationId": "92af88e390872c664a6dc58dfb6ea74d",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TurbineCollection"
                                }
                            }
                        }
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerError"
                    }
                }
            }
        },
        "/api/farms/{farmId}/turbines": {
            "get": {
                "summary": "List all turbines for a specific farm",
                "operationId": "255743b18bc48ca137f084175796146b",
                "parameters": [
                    {
                        "name": "farmId",
                        "in": "path",
                        "description": "The ID of the farm to filter by",
                        "required": true,
                        "example": 5
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TurbineCollection"
                                }
                            }
                        }
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerError"
                    }
                }
            }
        },
        "/api/turbines/{turbineId}": {
            "get": {
                "summary": "Show turbine details",
                "operationId": "9ef9e30ce72a62a38537f141054d9851",
                "parameters": [
                    {
                        "name": "turbineId",
                        "in": "path",
                        "description": "The ID of the turbine",
                        "required": true,
                        "example": 5
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TurbineResource"
                                }
                            }
                        }
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerError"
                    }
                }
            }
        },
        "/api/farms/{farmId}/turbines/{turbineId}": {
            "get": {
                "summary": "Show turbine details related to a specific farm",
                "operationId": "4b30229f78101355701a6f61ac7286c0",
                "parameters": [
                    {
                        "name": "farmId",
                        "in": "path",
                        "description": "The ID of the farm to filter by",
                        "required": true,
                        "example": 5
                    },
                    {
                        "name": "turbineId",
                        "in": "path",
                        "description": "The ID of the turbine",
                        "required": true,
                        "example": 5
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TurbineResource"
                                }
                            }
                        }
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerError"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "FarmCollection": {
                "description": "List farms",
                "properties": {
                    "data": {
                        "description": "Array of all farm data",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/FarmResource"
                        }
                    }
                },
                "type": "object"
            },
            "FarmResource": {
                "description": "Show farm details",
                "properties": {
                    "id": {
                        "description": "ID of the farm",
                        "type": "integer",
                        "example": 4
                    },
                    "name": {
                        "description": "Name of the farm",
                        "type": "integer",
                        "example": "Sample Farm"
                    },
                    "created_at": {
                        "description": "The datetime the farm was created",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-23T12:00:20.000000Z"
                    },
                    "updated_at": {
                        "description": "The datetime the farm was created",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-23T12:00:20.000000Z"
                    }
                },
                "type": "object"
            },
            "TurbineCollection": {
                "description": "List turbines",
                "properties": {
                    "data": {
                        "description": "Array of all turbine data",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/TurbineResource"
                        }
                    }
                },
                "type": "object"
            },
            "TurbineResource": {
                "description": "Show turbine details",
                "properties": {
                    "id": {
                        "description": "ID of the turbine",
                        "type": "integer",
                        "example": 4
                    },
                    "farm_id": {
                        "description": "ID of the parent farm",
                        "type": "integer",
                        "example": 3
                    },
                    "name": {
                        "description": "Name of the turbine",
                        "type": "integer",
                        "example": "Sample Turbine"
                    },
                    "created_at": {
                        "description": "The datetime the turbine was created",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-23T12:00:20.000000Z"
                    },
                    "updated_at": {
                        "description": "The datetime the turbine was created",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-23T12:00:20.000000Z"
                    }
                },
                "type": "object"
            }
        },
        "responses": {
            "InternalServerError": {
                "description": "Internal Server Error"
            }
        }
    }
}